requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
pillow==11.0.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.1
uritemplate==4.1.1
